Added gridplot function to RateCollection
The RateCollection class now is capable of plotting nuclides as cells in
a grid (Z vs. N), colored by mass fraction, molar abundances, or the
time derivatives of those variables. The plot is highly configurable
through keyword arguments.

Also added functions for evaluating ydot and activity for each nucleus,
where the latter is the sum over all of the terms contributing to ydot,
without taking sign into account (whether the nucleus is a reactant or
product). Neither of them currently consider screening.